Made CoreMemberMarshall an inner class of CoreMember to preserve the Class+inner marshal that other other marshable classes use.
